more systems, dod, imgui

+++ src/components/AnimationComponent.h
@@ -7,8 +7,8 @@ struct AnimationComponent{
     int currFrame;
     int frameSpeed;
     bool isLoop;
-    int startTime;
-    
+    Uint64 startTime;
+
     AnimationComponent(int frameNum = 1, int frameSpeed = 1, bool isLoop = true)
         : frameNum(frameNum), currFrame(1), frameSpeed(frameSpeed), isLoop(isLoop), startTime(SDL_GetTicks())
     {}
