+++ src/components/ProjectileComponent.h
@@ -5,10 +5,10 @@
 struct ProjectileComponent {  // change the whole projectile-health-damage system later
     bool isFriendly;
     int hitDamage;
-    int duration;
-    int startTime;
+    Uint64 duration;
+    Uint64 startTime;
 
-    ProjectileComponent(bool isFriendly = false, int hitDamage = 0, int duration = 0)
+    ProjectileComponent(bool isFriendly = false, int hitDamage = 0, Uint64 duration = 0)
         : isFriendly(isFriendly), hitDamage(hitDamage), duration(duration), startTime(SDL_GetTicks())
     {}
 };
