+++ src/components/SpriteComponent.h
@@ -11,8 +11,8 @@ struct SpriteComponent{
     bool isFixed;
     SDL_FRect srcRect;
 
-    SpriteComponent(std::string name = "", int width = 0, int height = 0, int zIndex = 0, bool isFixed = false, double srcRectX = 0, double srcRectY = 0)
+    SpriteComponent(std::string name = "", int width = 0, int height = 0, int zIndex = 0, bool isFixed = false, float srcRectX = 0.0f, float srcRectY = 0.0f)
         : name(std::move(name)), width(width), height(height), zIndex(zIndex), isFixed(isFixed),
-          srcRect{static_cast<float>(srcRectX), static_cast<float>(srcRectY), static_cast<float>(width), static_cast<float>(height)}
+          srcRect{srcRectX, srcRectY, static_cast<float>(width), static_cast<float>(height)}
     {}
 };
