+++ src/components/TransformComponent.h
@@ -5,9 +5,9 @@
 struct TransformComponent{
 	glm::vec2 position;
 	glm::vec2 scale;
-	double rotation;
+	float rotation;
 
-	TransformComponent(glm::vec2 position = glm::vec2(0.0, 0.0), glm::vec2 scale = glm::vec2(1.0, 1.0), double rotation = 0)
+	TransformComponent(glm::vec2 position = glm::vec2(0.0f, 0.0f), glm::vec2 scale = glm::vec2(1.0f, 1.0f), float rotation = 0.0f)
 		: position(position), scale(scale), rotation(rotation)
 	{}
 };
